corrige tipo dos parâmetros no Entidade::setPosicao

+++ src/entidades/Entidade.cpp
@@ -3,17 +3,17 @@
 namespace Jogo::Entidades {
 // Entidade::Entidade() ;
 // Entidade::~Entidade() = default;
-void Entidade::setPosicao(const int posX, const int posY) {
+void Entidade::setPosicao(const float posX, const float posY) {
   x = posX;
   y = posY;
 
-  pFig->setPosition((float)x, (float)y);
+  pFig->setPosition(x, y);
 }
 
-void Entidade::setPosicao(const sf::Vector2<int> &pos) {
+void Entidade::setPosicao(const sf::Vector2<float> &pos) {
   x = pos.x;
   y = pos.y;
 
-  pFig->setPosition((float)x, (float)y);
+  pFig->setPosition(x, y);
 }
 } // namespace Jogo::Entidades
