+++ src/entidades/Entidade.h
@@ -8,8 +8,8 @@ public:
   Entidade() = default;
   virtual ~Entidade() = default;
 
-  void setPosicao(const int x, const int y);
-  void setPosicao(const sf::Vector2<int> &);
+  void setPosicao(const float x, const float y);
+  void setPosicao(const sf::Vector2<float> &);
 
 protected:
   float x, y;
